requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.8
django-cors-headers==3.4.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
Naked==0.1.31
psycopg2-binary==2.8.5
pycparser==2.20
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
gunicorn==19.7.1
whitenoise==5.0.1